Use correct types

+++ apps/Arm/vm_qemu_virtio/src/cross_vm_connection.c
@@ -39,7 +39,7 @@ static int consume_callback(vm_t *vm, void *cookie)
     return 0;
 }
 
-extern unsigned long linux_ram_base;
+extern uintptr_t linux_ram_base;
 
 /* VM1 does not define this */
 seL4_Word WEAK intervm_sink_notification_badge(void);
